+++ cs.js
@@ -1146,10 +1146,20 @@ export default {
   'payment.option.notFount': 'You don\'t have any payment option saved.',
   'payment.option.create': 'Add a new payment option',
   'payment.option.description': 'Here is a list of your saved payment methods. You can save new payment methods here for faster checkout. Literary Universe does not store nor process your payment credentials.',
-  'payment.option.blueSnap': 'Blue Snap is our payment processor.',
+  'payment.option.blueSnap': 'Stripe is our payment processor.',
   'payment.history.title': 'Payment history',
   'payment.history.tab': 'History',
   'payment.history.none': 'We don\'t have any records of past payments.',
+  // Added 2020-04-30
+  'pm.new.aria': 'New conversation',
+  'pm.new.buttonTitle': 'Start a new conversation',
+  'plans.options.aria': 'Payment options',
+  'plans.options.creditCard': 'Credit Card',
+  'plans.purchase.action': 'Subscribe',
+  'plans.purchase.title': 'Buy Literary Universe subscription',
+  'payment.creditCard.details': 'Credit Card details',
+  'payments.zipcode.label': 'ZIP code',
+  'form.validator.zipCode': 'Must be a valid ZIP code',
 }
 
 /* eslint-enable max-len */